requirements.txt
langchain>=0.2.14
langchain-community>=0.2.12
langchain-openai>=0.1.20
langchain-chroma>=0.1.2
langgraph>=0.2.35
chromadb>=0.5.5
pypdf>=4.2.0
streamlit>=1.36.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
Ipython
python-dotenv
langchain-text-splitters